Adding a service to run on all nodes and keep the local images on each node in sync with the running experiment services

+++ start_registry_container.sh
@@ -55,20 +55,29 @@ docker service create -d \
    # omitting the frinexDockerRegistry volume for the service because expecting it to exist also requires it to be synchronised across all nodes
    # -v frinexDockerRegistry:/var/lib/registry \
 
-# build the frinex_synchronisation_service
-docker build --rm -f docker/frinex_synchronisation_service.Dockerfile -t DOCKER_REGISTRY/frinex_synchronisation_service:latest .
-docker push DOCKER_REGISTRY/frinex_synchronisation_service:latest
+cd $(dirname "$0")
+workingDir=$(pwd -P)
+cd $(dirname "$0")/src/main/
 
-read -p "Press enter to restart frinex_synchronisation_service"
-# remove the old frinex_synchronisation_service
-docker service rm frinex_synchronisation_service 
-# start the frinex_synchronisation_service
-docker service create -e 'ServiceHostname={{.Node.Hostname}}' --cpus=".5" --mount=type=bind,src=/var/run/docker.sock,dst=/var/run/docker.sock -dit --name frinex_synchronisation_service DOCKER_REGISTRY/frinex_synchronisation_service:latest
+# check that the properties to be used match the current machine
+if ! grep -q $(hostname) config/publish.properties; then 
+    echo "Aborting because the publish.properties does not match the current machine.";
+else
+   # build the frinex_synchronisation_service
+   docker build --rm -f docker/frinex_synchronisation_service.Dockerfile -t DOCKER_REGISTRY/frinex_synchronisation_service:latest .
+   docker push DOCKER_REGISTRY/frinex_synchronisation_service:latest
 
-# for currentService in $(sudo docker service ls | grep -E "_staging|_production" | grep -E "_admin|_web" | awk '{print $2}')
-# do
-#    echo $currentService
-#    # push each web and admin service image currently in use to the empty registry so that they can be accessed by the swarm nodes
-#    # sudo docker push $currentService
-#    # curl "http://frinexbuild:8010/cgi/frinex_restart_experient.cgi?$currentService"
-# done
+   read -p "Press enter to restart frinex_synchronisation_service"
+   # remove the old frinex_synchronisation_service
+   docker service rm frinex_synchronisation_service 
+   # start the frinex_synchronisation_service
+   docker service create -e 'ServiceHostname={{.Node.Hostname}}' --cpus=".5" --mount=type=bind,src=/var/run/docker.sock,dst=/var/run/docker.sock -dit --name frinex_synchronisation_service DOCKER_REGISTRY/frinex_synchronisation_service:latest
+
+   # for currentService in $(sudo docker service ls | grep -E "_staging|_production" | grep -E "_admin|_web" | awk '{print $2}')
+   # do
+   #    echo $currentService
+   #    # push each web and admin service image currently in use to the empty registry so that they can be accessed by the swarm nodes
+   #    # sudo docker push $currentService
+   #    # curl "http://frinexbuild:8010/cgi/frinex_restart_experient.cgi?$currentService"
+   # done
+fi
